Debugging db errors

+++ BusinessApp/UI/ManageCustomers.cs
@@ -20,6 +20,7 @@ namespace SSC.UI
         private Navbar aNavbar;
         private Admin Admin;
         DataTable dt = new DataTable();
+        private int selectedRow ;
         public ManageCustomers()
         {
             InitializeComponent();
@@ -81,9 +82,6 @@ namespace SSC.UI
                 case "manageTables":
                     OpenForm(new ManageTables(this.Size, this.Location, Admin));
                     break;
-                case "feedback":
-                    OpenForm(new ManageFeedback(this.Size, this.Location, Admin));
-                    break;
                 default:
                     break;
             }
@@ -111,7 +109,6 @@ namespace SSC.UI
                 aNavbar.BringToFront();
             }
         }
-
         private void menuGridView_CellContentClick(object sender, DataGridViewCellEventArgs e)
         {
 
@@ -235,24 +232,39 @@ namespace SSC.UI
 
         private void deleteBtn_Click(object sender, EventArgs e)
         {
-            if (dataGridView1.SelectedRows.Count > 0)
+            DialogResult dialogResult = MessageBox.Show("Are you sure you want to delete this customer?", "Delete Customer", MessageBoxButtons.YesNo);
+            if (dialogResult == DialogResult.Yes)
             {
-                try
+                if (dataGridView1.SelectedRows.Count > 0)
                 {
-                    int id = Convert.ToInt32(dataGridView1.SelectedRows[0].Cells["ID"].Value);
-                    string status = dataGridView1.SelectedRows[0].Cells["Status"].Value.ToString().ToLower();
-                    dt.Rows.Clear();
-                    ObjectHandler.GetCustomerDL().DeleteCustomer(Convert.ToInt32(id), status, ObjectHandler.GetUserDL().GetUserID(id));
-                    LoadData();
+                    try
+                    {
+                        selectedRow = dataGridView1.SelectedRows[0].Index;
+
+                        if (dataGridView1 != null && selectedRow >= 0 && selectedRow < dataGridView1.Rows.Count)
+                        {
+                            DataGridViewRow selectedDataGridViewRow = dataGridView1.Rows[selectedRow];
+
+                            if (selectedDataGridViewRow != null && selectedDataGridViewRow.Cells["ID"].Value != null)
+                            {
+                                int id = Convert.ToInt32(dataGridView1.SelectedRows[0].Cells["ID"].Value);
+                                string status = dataGridView1.SelectedRows[0].Cells["Status"].Value.ToString().ToLower();
+                                dt.Rows.Clear();
+                                ObjectHandler.GetCustomerDL().DeleteCustomer(Convert.ToInt32(id), status, ObjectHandler.GetUserDL().GetUserID(id));
+                                LoadData();
+                            }
+                        }
+                    }
+                    catch (Exception ex)
+                    {
+                        MessageBox.Show(ex.ToString());
+                    }
                 }
-                catch (Exception ex)
+                else
                 {
-                    MessageBox.Show(ex.ToString());
+                    MessageBox.Show("Please select a row to delete");
                 }
-            }
-            else
-            {
-                MessageBox.Show("Please select a row to delete");
+
             }
         }
     }
